Abstract data to own file
